pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "comment-vision-claw"
version = "0.1.0"
description = "抖音热评抓取与分析工具 - 自动捕获高赞评论并生成分析报告"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Comment-Vision-Claw Contributors"}
]
keywords = ["douyin", "tiktok", "comments", "analysis", "mcp", "social-media"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dependencies = [
    "python-dotenv>=1.0.0",
    "playwright>=1.40.0",
    "playwright-stealth>=1.0.6",
    "mcp>=0.9.0",
    "openai>=1.0.0",  # Optional, for AI analysis
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
comment-vision-claw = "server:cli_main"
cvc = "server:cli_main"  # Short alias

[project.urls]
Homepage = "https://github.com/wp-i/comment-vision-claw"
Repository = "https://github.com/wp-i/comment-vision-claw"
Issues = "https://github.com/wp-i/comment-vision-claw/issues"

[tool.setuptools.packages.find]
include = ["engine*", "prompts*"]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
